Fully implemented ActionBar. Changed complete order for more efficiency.
The plugin only handles positive exp values.
All if statements which include an and were included in the same
statement instead.
There can be a feature to get experience as soon as you finish healing.
Interesting?
Changed the display so plugin no longer puts exp when player is full HP.